==> 1. Procesado de datos.ipynb <==
### 1. PROCESADO DE DATOS

#Librerias
import pandas as pd
import numpy as np
import os

#Rutas
path = 'Data/'
footballpath = 'Data/football/ligasprincipales/'
footballotherspath = 'Data/football/otrasligas/'
basketballpath = 'Data/basketball/'
tennismenpath = 'Data/tennis/men/'
tenniswomenpath = 'Data/tennis/women/'

#Columnas
footballcolumns = ['Div', 'Date', 'HomeTeam', 'AwayTeam', 'FTHG', 'FTAG', 'FTR', 'IWH', 'IWA']
tennismencolumns = ['ATP', 'Location', 'Tournament', 'Date', 'Series', 'Court', 'Surface','Round', 'Best of', 'Winner', 'Loser', 'B365W', 'B365L']
tenniswomencolumns = ['WTA', 'Location', 'Tournament', 'Date', 'Tier', 'Court', 'Surface','Round', 'Best of', 'Winner', 'Loser', 'B365W', 'B365L']
footballotherscolumns =['Country', 'League', 'Season', 'Date', 'Time', 'Home', 'Away', 'HG','AG', 'Res', 'PH', 'PD', 'PA']

#Nuevo Directorio
os.makedirs("Data/Depurado")

#Files
footballfiles = os.listdir(path=footballpath)
footballothersfiles = os.listdir(path=footballotherspath)
basketballfiles = os.listdir(path=basketballpath)
tennismenfiles = os.listdir(path=tennismenpath)
tenniswomenfiles = os.listdir(path=tenniswomenpath)

#Football
vacio = pd.DataFrame()

for file in footballfiles:
    tempf1 = pd.read_excel(footballpath+file, sheet_name=None)
    sheets = list(tempf1.keys())
    for sheet in sheets:
        tempf2 = pd.read_excel(footballpath+file, sheet_name=sheet)
        tempf2 = tempf2[footballcolumns]
        vacio = vacio.append(tempf2)
        
vacio = vacio.drop_duplicates(keep=False)
vacio = vacio.dropna(how='any')

vacio.to_excel('Data/Depurado/football.xlsx')

#Football Others
vacio = pd.DataFrame()

for file in footballothersfiles:
    tempfo1 = pd.read_excel(footballotherspath+file, sheet_name=None)
    sheets = list(tempfo1.keys())
    for sheet in sheets:
        tempfo2 = pd.read_excel(footballotherspath+file, sheet_name=sheet)
        tempfo2 = tempfo2[footballotherscolumns]
        vacio = vacio.append(tempfo2)
        
vacio = vacio.drop_duplicates(keep=False)
vacio = vacio.dropna(how='any')

vacio.to_excel('Data/Depurado/footballothers.xlsx')

#Basketball
vacio = pd.DataFrame()

for file in basketballfiles:
    tempb1 = pd.read_excel(basketballpath+file, sheet_name=None)
    sheets = list(tempb1.keys())
    for sheet in sheets:
        tempb2 = pd.read_excel(basketballpath+file, sheet_name=sheet)
        vacio = vacio.append(tempb2)
        
vacio = vacio.drop_duplicates(keep=False)
vacio = vacio.dropna(how='any')
        
vacio.to_excel('Data/Depurado/basketball.xlsx')

#TennisMen
vacio = pd.DataFrame()

for file in tennismenfiles:
    temptm1 = pd.read_excel(tennismenpath+file, sheet_name=None)
    sheets = list(temptm1.keys())
    for sheet in sheets:
        temptm2 = pd.read_excel(tennismenpath+file, sheet_name=sheet)
        temptm2 = temptm2[tennismencolumns]
        vacio = vacio.append(temptm2)
        
vacio = vacio.drop_duplicates(keep=False)
vacio = vacio.dropna(how='any')
        
vacio.to_excel('Data/Depurado/tennismen.xlsx')

#TennisWomen
vacio = pd.DataFrame()

for file in tenniswomenfiles:
    temptw1 = pd.read_excel(tenniswomenpath+file, sheet_name=None)
    sheets = list(temptw1.keys())
    for sheet in sheets:
        temptw2 = pd.read_excel(tenniswomenpath+file, sheet_name=sheet)
        temptw2 = temptw2[tenniswomencolumns]
        vacio = vacio.append(temptw2)
        
vacio = vacio.drop_duplicates(keep=False)
vacio = vacio.dropna(how='any')
        
vacio.to_excel('Data/Depurado/tenniswomen.xlsx')

#Explorar Football

football = pd.read_excel('Data/Depurado/football.xlsx')
football.head(5)

#Mirando los tipos del dataframe
football.dtypes

#Explorar Football Others

footballothers = pd.read_excel('Data/Depurado/footballothers.xlsx')
footballothers.head(5)

#Mirando los tipos del dataframe
footballothers.dtypes

#Explorar Basketball

basketball = pd.read_excel('Data/Depurado/basketball.xlsx')
basketball.head(5)

#Mirando los tipos del dataframe
basketball.dtypes

#Explorar TennisMen

tennismen = pd.read_excel('Data/Depurado/tennismen.xlsx')
tennismen.head(5)

#Mirando los tipos del dataframe
tennismen.dtypes

#Explorar TennisWomen

tenniswomen = pd.read_excel('Data/Depurado/tenniswomen.xlsx')
tenniswomen.head(5)

#Mirando los tipos del dataframe
tenniswomen.dtypes

#Juntar Tennis masculino con Tennis femenino con las columnas necesarias

tennisall = tennismen.append(tenniswomen)
tennisall = tennisall[['Tournament', 'Date','Winner', 'B365W', 'B365L']]
tennisall['Sport'] = 'Tennis'

tennisall.head()

tennisall.dtypes

#Convertir tipo de una columna a float

tennisall['B365L'] = tennisall['B365L'].convert_objects(convert_numeric=True)

#Determinar si gana el favorito

tennisall['GanaFavorito'] = tennisall['B365W'] < tennisall['B365L']
tennisall.loc[tennisall['GanaFavorito'] == False, 'GanaFavorito'] = 0
tennisall.loc[tennisall['GanaFavorito'] == True, 'GanaFavorito'] = 1

#Cambiar nombre las columnas

tennisall = tennisall.rename(columns={"B365W":"OddW", "B365L":"OddL",'Winner':'Favorito'})
tennisall['CuotaFavorito'] = tennisall[['OddW','OddL']].min(axis=1)

tennisall = tennisall[['Tournament','Date','Sport','Favorito','GanaFavorito','CuotaFavorito']]

tennisall.head()

#Basketball

basketball = basketball[['Date','League','Winner','H365','A365']]
basketball['Sport'] = 'Basketball'

basketball.head()

#Determinando cual es el equipo favorito, será el equipo que menos cuota tiene.
basketball['Favorito'] = basketball['H365'] > basketball['A365']
basketball.loc[basketball['Favorito'] == False, 'Favorito'] = 'H'
basketball.loc[basketball['Favorito'] == True, 'Favorito'] = 'A'

basketball['GanaFavorito'] = basketball['Winner'] == basketball['Favorito']
basketball.loc[basketball['GanaFavorito'] == False, 'GanaFavorito'] = 0
basketball.loc[basketball['GanaFavorito'] == True, 'GanaFavorito'] = 1

basketball['CuotaFavorito'] = basketball[['H365','A365']].min(axis=1)

basketball = basketball.rename(columns={"League":"Tournament"})

basketball = basketball[['Date','Tournament','Sport','Favorito','GanaFavorito','CuotaFavorito']]

basketball.head()

#Football

football = football[['Div','Date','FTR','IWH','IWA']]
football['Sport'] = 'Football'

#Determinando cual es el equipo favorito, será el equipo que menos cuota tiene.
football['Favorito'] = football['IWH'] > football['IWA']

football.loc[football['Favorito'] == False, 'Favorito'] = 'H'
football.loc[football['Favorito'] == True, 'Favorito'] = 'A'

football['GanaFavorito'] = football['FTR'] == football['Favorito']
football.loc[football['GanaFavorito'] == False, 'GanaFavorito'] = 0
football.loc[football['GanaFavorito'] == True, 'GanaFavorito'] = 1

football['CuotaFavorito'] = football[['IWH','IWA']].min(axis=1)

football = football.rename(columns={"Div":"Tournament"})

football = football[['Tournament','Date','Sport','Favorito','GanaFavorito','CuotaFavorito']]

football.head(10)

#Football Others

footballothers = footballothers[['Country','Date','Res','PH','PA']]
footballothers['Sport'] = 'Football'

footballothers.head()

#Determinando cual es el equipo favorito, será el equipo que menos cuota tiene.
footballothers['Favorito'] = footballothers['PH'] > footballothers['PA']
footballothers.loc[footballothers['Favorito'] == False, 'Favorito'] = 'H'
footballothers.loc[footballothers['Favorito'] == True, 'Favorito'] = 'A'

footballothers['GanaFavorito'] = footballothers['Res'] == footballothers['Favorito']
footballothers.loc[footballothers['GanaFavorito'] == False, 'GanaFavorito'] = 0
footballothers.loc[footballothers['GanaFavorito'] == True, 'GanaFavorito'] = 1

footballothers['CuotaFavorito'] = footballothers[['PH','PA']].min(axis=1)

footballothers = footballothers.rename(columns={"Country":"Tournament"})

footballothers = footballothers[['Tournament','Date','Sport','Favorito','GanaFavorito','CuotaFavorito']]
footballothers.head()

#Juntaremos todos los dataframes en uno para posteriormente guararlo en un archivo Excel.
dataframes = (football,footballothers,tennisall,basketball)

sports = pd.concat(dataframes)

sports

#Guardamos el dataframe en excel para utilizarlo en el siguiente notebook
sports.to_excel(path+'sports.xlsx')
