==> 2. Hucha.ipynb <==
### 2. HUCHA

#Librerias
import pandas as pd
import numpy as np
import os

#Creamos la ruta
path = 'Data/'
sports = pd.read_excel(path+'sports.xlsx')

#Observamos las primeras filas del dataframe
sports.head()

#Creamos una función que convierta la cuota en una probabilidad, para así hacer un análisis más sencillo.
#Esta probabilidad será la que tiene de ganar el favorito

def prob(a):
    return (1/a)*100

#Aplicamos la formula en una nueva columna
sports['Probabilidad'] = sports['CuotaFavorito'].apply(prob)

sports.head()

sports.dtypes

## El primer análisis que haremos será apostar 10€ siempre al favorito.

#Si ganamos tendremos el beneficio menos lo apostado, para saber el beneficio neto y si perdemos restaremos lo apostado.
sports['Hucha10'] = np.where(sports['GanaFavorito'] == 1, sports['CuotaFavorito']*10-10, -10)

sports.head()

#Sumaremos todo para saber si hubieramos ganado o perdido dinero.
sports['Hucha10'].sum()

#La segunda Hucha la calcularemos multiplicando su cuota por la probabilidad de ganar
sports['HuchaB'] = np.where(sports['GanaFavorito'] == 1, (sports['CuotaFavorito']*sports['Probabilidad']/10)-(sports['Probabilidad']/10), -sports['Probabilidad']/10)

sports['HuchaB'].sum()

La hucha sale negativa, entonces tendremos que buscar un indicador que nos diga cuanto apostar a cada cuota.

#Cogemos solamente los favoritos que ganan

favoritos = sports[sports['GanaFavorito'] == 1]
favoritos.head()

#Agrupado por cuotas

favoritoAgrupados = favoritos.groupby('CuotaFavorito',as_index=False)['GanaFavorito'].sum()
favoritoAgrupados.head()

sportsTotal = sports.groupby('CuotaFavorito',as_index=False)['Favorito'].count()
sportsTotal.head()

#Juntamos las tablas para poder obtener el porcentaje de favoritos que ganan teniendo una cuota concreta
sportsTotalFavoritos = favoritoAgrupados.merge(sportsTotal, how='left')
sportsTotalFavoritos['Porcentaje'] = sportsTotalFavoritos['GanaFavorito']/sportsTotalFavoritos['Favorito']
sportsTotalFavoritos['Probabilidad'] = sportsTotalFavoritos['CuotaFavorito'].apply(prob)
sportsTotalFavoritos['Indicador'] = sportsTotalFavoritos['Porcentaje'] * sportsTotalFavoritos['Probabilidad']
sportsTotalFavoritos = sportsTotalFavoritos[['CuotaFavorito','Indicador']]
sportsTotalFavoritos.head()

#Guardaremos los indicadores en un archivo
sportsTotalFavoritos.to_csv(path+'indicadores.csv')

#Los indicadores los juntaremos en un único datafram
sports = sports.merge(sportsTotalFavoritos, how = 'left')

sports.head()

sports['HuchaC'] = np.where(sports['GanaFavorito'] == 1, (sports['CuotaFavorito']*sports['Indicador']/10)-sports['Indicador']/10, -sports['Indicador']/10)

sports.head()

print(sports['Hucha10'].sum())
print(sports['HuchaB'].sum())
print(sports['HuchaC'].sum())

#Gracias a las indicador hemos reducido las perdidas pero seguimos en negativo.

Ahora vamos a calcular las tres huchas para Fútbol, Baloncesto y Tenis.

sports.groupby('Sport', as_index=False)[['Hucha10','HuchaB','HuchaC']].sum()

Para finalizar vamos a analizar los 10 torneos de fútbol que son más rentables si apostamos al favorito.

footballTournament = sports[sports['Sport'] == 'Football'].groupby('Tournament', as_index=False)[['Hucha10','HuchaB','HuchaC']].sum()
footballTournament.sort_values(by='HuchaC', ascending=False).head(10)

# Estadística descriptiva

datos = sports[['CuotaFavorito','Probabilidad','HuchaC','Sport']]

#### Datos estadísticos globales

datos.describe()

datos.mode()

#### Datos estadísticos Fútbol

datos[datos['Sport'] == 'Football'].describe()

datos[datos['Sport'] == 'Football'].mode()

datos[datos['Sport'] == 'Football'].median()

#### Datos estadísticos Baloncesto

datos[datos['Sport'] == 'Basketball'].describe()

datos[datos['Sport'] == 'Basketball'].mode()

datos[datos['Sport'] == 'Basketball'].median()

#### Datos estadísticos Tenis

datos[datos['Sport'] == 'Tennis'].describe()

datos[datos['Sport'] == 'Tennis'].mode()

datos[datos['Sport'] == 'Tennis'].median()

## Creamos un fichero excel para utilizar en Rstudio

sports.to_csv(path+'visualizar_rstudio.csv')
